+++ packages/frame3d/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/frame3d",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "THREE.js 3D rendering frame in which DomElement 3D can be added",
   "main": "src",
   "scripts": {
@@ -20,7 +20,7 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/frame3d",
   "peerDependencies": {
-    "@ud-viz/utils_browser": "4.0.3",
+    "@ud-viz/utils_browser": "4.1.0",
     "itowns": "2.41.1-next.9",
     "three": "0.154.0"
   }
