+++ packages/game_browser_template/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/game_browser_template",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "Collection of script base on @ud-viz/game_browser",
   "main": "src",
   "scripts": {
@@ -20,11 +20,11 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/game_browser_template",
   "peerDependencies": {
-    "@ud-viz/game_browser": "4.0.3",
-    "@ud-viz/game_shared": "4.0.3",
-    "@ud-viz/game_shared_template": "4.0.3",
-    "@ud-viz/utils_browser": "4.0.3",
-    "@ud-viz/utils_shared": "4.0.3",
+    "@ud-viz/game_browser": "4.1.0",
+    "@ud-viz/game_shared": "4.1.0",
+    "@ud-viz/game_shared_template": "4.1.0",
+    "@ud-viz/utils_browser": "4.1.0",
+    "@ud-viz/utils_shared": "4.1.0",
     "itowns": "2.41.1-next.9",
     "three": "0.154.0"
   }
