+++ packages/game_editor/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/game_editor",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "game editor",
   "main": "src",
   "scripts": {
@@ -20,12 +20,12 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/game_editor",
   "peerDependencies": {
-    "@ud-viz/game_shared": "4.0.3",
-    "@ud-viz/game_shared_template": "4.0.3",
-    "@ud-viz/game_browser": "4.0.3",
-    "@ud-viz/game_browser_template": "4.0.3",
-    "@ud-viz/utils_browser": "4.0.3",
-    "@ud-viz/utils_shared": "4.0.3",
+    "@ud-viz/game_shared": "4.1.0",
+    "@ud-viz/game_shared_template": "4.1.0",
+    "@ud-viz/game_browser": "4.1.0",
+    "@ud-viz/game_browser_template": "4.1.0",
+    "@ud-viz/utils_browser": "4.1.0",
+    "@ud-viz/utils_shared": "4.1.0",
     "three": "0.154.0"
   }
 }
