+++ packages/game_shared_template/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/game_shared_template",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "Collection of script base on @ud-viz/game_shared",
   "main": "src",
   "scripts": {
@@ -20,8 +20,8 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/game_shared_template",
   "peerDependencies": {
-    "@ud-viz/game_shared": "4.0.3",
-    "@ud-viz/utils_shared": "4.0.3",
+    "@ud-viz/game_shared": "4.1.0",
+    "@ud-viz/utils_shared": "4.1.0",
     "three": "0.154.0"
   }
 }
