+++ packages/point_cloud_visualizer/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/point_cloud_visualizer",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "application to visualize a 3DTiles point cloud",
   "main": "src",
   "scripts": {
@@ -20,8 +20,8 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/point_cloud_visualizer",
   "peerDependencies": {
-    "@ud-viz/utils_browser": "4.0.3",
-    "@ud-viz/utils_shared": "4.0.3",
+    "@ud-viz/utils_browser": "4.1.0",
+    "@ud-viz/utils_shared": "4.1.0",
     "itowns": "2.41.1-next.9",
     "three": "0.154.0"
   }
