+++ packages/show_room/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/show_room",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "application using all @ud-viz packages",
   "main": "src",
   "scripts": {
@@ -20,21 +20,21 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/show_room",
   "peerDependencies": {
-    "@ud-viz/frame3d": "4.0.3",
-    "@ud-viz/smdb": "4.0.3",
-    "@ud-viz/utils_browser": "4.0.3",
-    "@ud-viz/widget_camera_positioner": "4.0.3",
-    "@ud-viz/widget_geocoding": "4.0.3",
-    "@ud-viz/widget_layer_choice": "4.0.3",
-    "@ud-viz/widget_slide_show": "4.0.3",
-    "@ud-viz/widget_sparql": "4.0.3",
-    "@ud-viz/widget_3d_tiles": "4.0.3",
-    "@ud-viz/widget_bookmark": "4.0.3",
-    "@ud-viz/game_browser": "4.0.3",
-    "@ud-viz/game_browser_template": "4.0.3",
-    "@ud-viz/game_shared": "4.0.3",
-    "@ud-viz/game_shared_template": "4.0.3",
-    "@ud-viz/widget_legonizer": "4.0.3",
+    "@ud-viz/frame3d": "4.1.0",
+    "@ud-viz/smdb": "4.1.0",
+    "@ud-viz/utils_browser": "4.1.0",
+    "@ud-viz/widget_camera_positioner": "4.1.0",
+    "@ud-viz/widget_geocoding": "4.1.0",
+    "@ud-viz/widget_layer_choice": "4.1.0",
+    "@ud-viz/widget_slide_show": "4.1.0",
+    "@ud-viz/widget_sparql": "4.1.0",
+    "@ud-viz/widget_3d_tiles": "4.1.0",
+    "@ud-viz/widget_bookmark": "4.1.0",
+    "@ud-viz/game_browser": "4.1.0",
+    "@ud-viz/game_browser_template": "4.1.0",
+    "@ud-viz/game_shared": "4.1.0",
+    "@ud-viz/game_shared_template": "4.1.0",
+    "@ud-viz/widget_legonizer": "4.1.0",
     "itowns": "2.41.1-next.9",
     "three": "0.154.0"
   }
