+++ packages/utils_node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/utils_node",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "utils methods interpretable in a nodejs",
   "main": "src",
   "scripts": {
@@ -23,7 +23,7 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/utils_node",
   "peerDependencies": {
-    "@ud-viz/utils_shared": "4.0.3",
+    "@ud-viz/utils_shared": "4.1.0",
     "child-process-promise": "2.2.1",
     "puppeteer": "19.4.1"
   }
