+++ packages/widget_bookmark/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/widget_bookmark",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "widget bookmark of the itowns framework",
   "main": "src",
   "scripts": {
