+++ packages/widget_camera_positioner/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/widget_camera_positioner",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "widget base map of the itowns framework",
   "main": "src",
   "scripts": {
