+++ packages/widget_legonizer/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/widget_legonizer",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "widget layer choice of the itowns framework",
   "main": "src",
   "scripts": {
@@ -20,8 +20,8 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/widget_legonizer",
   "peerDependencies": {
-    "@ud-viz/game_browser": "4.0.3",
-    "@ud-viz/utils_browser": "4.0.3",
+    "@ud-viz/game_browser": "4.1.0",
+    "@ud-viz/utils_browser": "4.1.0",
     "legonizer": "^1.0.0",
     "itowns": "2.41.1-next.9",
     "three": "0.154.0"
