+++ packages/widget_slide_show/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/widget_slide_show",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "widget slide show of the itowns framework",
   "main": "src",
   "scripts": {
@@ -20,7 +20,7 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/widget_slide_show",
   "peerDependencies": {
-    "@ud-viz/utils_browser": "4.0.3",
+    "@ud-viz/utils_browser": "4.1.0",
     "itowns": "2.41.1-next.9",
     "three": "0.154.0"
   }
