+++ packages/widget_sparql/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ud-viz/widget_sparql",
-  "version": "4.0.3",
+  "version": "4.1.0",
   "description": "widget sparql of the itowns framework",
   "main": "src",
   "scripts": {
@@ -20,7 +20,7 @@
   },
   "homepage": "https://github.com/VCityTeam/UD-Viz/packages/widget_sparql",
   "peerDependencies": {
-    "@ud-viz/utils_browser": "4.0.3",
+    "@ud-viz/utils_browser": "4.1.0",
     "three": "0.154.0",
     "d3": "^7.8.5"
   }
